+++ common/mouse_temponodes.c
@@ -42,7 +42,7 @@ extern struct Root *root;
 
 int MoveTempoNode_Mouse(
 	struct Tracker_Windows *window,
-	int x,int y
+	float x,float y
 ){
 	struct WBlocks *wblock=window->wblock;
 	struct MouseAction *action= &window->prevaction;
