+++ common/mouse_vellinenode.c
@@ -36,7 +36,7 @@ extern struct Root *root;
 
 int MoveVelNode_Mouse(
 	struct Tracker_Windows *window,
-	int x,int y
+	float x,float y
 ){
 	struct WBlocks *wblock=window->wblock;
 	struct MouseAction *action= &window->prevaction;
