+++ common/mouse_vellinestart.c
@@ -39,7 +39,7 @@ extern struct Root *root;
 
 int MoveVelStart_Mouse(
 	struct Tracker_Windows *window,
-	int x,int y
+	float x,float y
 ){
 	struct WBlocks *wblock=window->wblock;
 	struct MouseAction *action= &window->prevaction;
